file_service/pyproject.toml
[tool.poetry]
name = "file-service"
version = "0.1.0"
description = ""
authors = ["Simon Sukovatitsin <simonsnag@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
minio = "^7.2.7"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
